# specs/001-captura-inventarios/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Captura Inteligente de Inventarios — API
  version: 0.1.0
  description: |
    Contrato derivado de los esquemas Zod en `packages/contracts` (D-22).
    Este documento se GENERA; no se edita a mano.

    Dos invariantes gobiernan el contrato:
    - **El saldo esperado nunca aparece en una respuesta dirigida a rol Operador**, ni directa
      ni derivable (FR-1.18, SC-1.3). Las alertas de discrepancia no exponen magnitud.
    - **Nada se actualiza.** No existe PUT ni PATCH sobre el libro; una corrección es un POST
      que añade un registro con mayor secuencia (D2, FR-1.13).

    Autenticación por cookie de sesión `httpOnly`; el servidor no guarda estado (D-20).
servers:
  - url: https://api.inventarios.example/v1
security:
  - sesionCookie: []
...
paths:
...
  /rondas:
    post:
      tags: [captura]
      summary: Abrir una ronda de conteo
      description: Crea una ronda propia del operador. No altera ninguna otra ronda (FR-1.5).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bodegaId, claveIdempotencia, desfaseRelojMs]
              properties:
                bodegaId: { type: string, format: uuid }
                claveIdempotencia: { type: string, format: uuid }
                desfaseRelojMs: { type: integer }
      responses:
        '201':
          description: Ronda abierta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ronda' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '409': { $ref: '#/components/responses/Conflicto' }
  /rondas/{rondaId}/catalogo:
    get:
      tags: [captura]
      summary: Catálogo de referencia de la bodega
      description: |
        Catálogo de artículos de la bodega. **No es una lista dirigida**: el Operador dicta lo que
        ve (D7), no sigue un orden propuesto. Se usa para el cuadre de cierre, para la
        búsqueda manual y como caché local de la resolución de nombres.
        **No incluye el saldo esperado** — ni el valor, ni un rango, ni nada que permita inferirlo.
      parameters:
        - { name: rondaId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limite, in: query, schema: { type: integer, default: 200, maximum: 500 } }
      responses:
        '200':
          description: Página del catálogo de la bodega
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ArticuloDeTrabajo' }
                  siguienteCursor: { type: string, nullable: true }
  /rondas/{rondaId}/resolucion-articulo:
    post:
      tags: [captura]
      summary: Resolver el nombre dictado contra el catálogo
      description: |
        Ruta caliente de cada turno (D-05, D-19). Recibe el texto del nombre tal como lo
        pronunció el Operador y lo resuelve por `pg_trgm` sobre nombres y alias de la bodega.

        Devuelve `resuelto` cuando el mejor candidato supera el umbral de aceptación **y** el
        margen contra el segundo es suficiente. En cualquier otro caso devuelve `candidatos`
        para que el Operador escoja tocando la pantalla: **el sistema nunca elige en caso de
        duda** (FR-1.11, FR-1.27).

        **No incluye el saldo esperado** de ningún candidato.
      parameters:
        - { name: rondaId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [textoDictado]
              properties:
                textoDictado: { type: string, minLength: 1, maxLength: 200 }
      responses:
        '200':
          description: Resolución del nombre
          content:
            application/json:
              schema:
                type: object
                required: [estado]
                properties:
                  estado: { type: string, enum: [resuelto, candidatos, sin_coincidencia] }
                  articulo:
                    allOf:
                      - $ref: '#/components/schemas/ArticuloDeTrabajo'
                    nullable: true
                    description: Presente solo si estado = resuelto.
                  origenNombre:
                    type: string
                    nullable: true
                    enum: [exacto, similitud, alias]
                  candidatos:
                    type: array
                    description: Presente si estado = candidatos. Ordenados por puntaje descendente.
                    items:
                      type: object
                      required: [articulo, puntaje]
                      properties:
                        articulo: { $ref: '#/components/schemas/ArticuloDeTrabajo' }
                        puntaje: { type: number, minimum: 0, maximum: 1 }
        '422': { $ref: '#/components/responses/EntradaInvalida' }
  /rondas/{rondaId}/registros:
    post:
      tags: [captura]
      summary: Registrar el conteo de un artículo
      description: |
        Append-only. Registrar dos veces el mismo artículo en la ronda **no suma**: añade un
        registro con mayor secuencia que supersede al anterior, y exige `confirmaCorreccion`
        (D4, FR-1.14). Deduplicación por `claveIdempotencia` (R2): reenviar es seguro.
      parameters:
        - { name: rondaId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegistroEntrada' }
      responses:
        '201':
          description: Registro aceptado. La validación llega por separado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistroAceptado' }
        '200':
          description: Ya existía un registro con esa clave de idempotencia. No se duplicó.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistroAceptado' }
        '409':
          description: El artículo ya fue contado en esta ronda y falta `confirmaCorreccion`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/EntradaInvalida' }
...
  /rondas/{rondaId}/cuadre-cierre:
    get:
      tags: [captura]
      summary: Artículos sin resolver antes de cerrar la ronda
      description: Lista lo no registrado, para decidir entre contado en cero y no contado (D1, FR-1.15).
      parameters:
        - { name: rondaId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Pendientes de decisión
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ArticuloDeTrabajo' }
  /rondas/{rondaId}/cierre:
    post:
      tags: [captura]
      summary: Cerrar la ronda
      description: |
        Falla si queda un artículo sin estado explícito (FR-1.15) o una alerta sin resolver
        (FR-2.8). Tras el cierre, los registros de la ronda son inmodificables (FR-1.17).
      parameters:
        - { name: rondaId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Ronda cerrada }
        '409':
          description: Quedan pendientes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    sesionCookie:
      type: apiKey
      in: cookie
      name: sesion
  schemas:
...
    ModoCaptura:
      type: string
      enum: [voz, texto]
    EstadoRegistro:
      type: string
      enum: [contado, contado_en_cero, no_contado]
...
    Ronda:
      type: object
      required: [id, bodegaId, operadorId, abiertaEn]
      properties:
        id: { type: string, format: uuid }
        bodegaId: { type: string, format: uuid }
        operadorId: { type: string, format: uuid }
        abiertaEn: { type: string, format: date-time }
    ArticuloDeTrabajo:
      type: object
      description: |
        Artículo tal como lo ve el Operador. Deliberadamente NO contiene saldo esperado,
        tolerancia, ni conteos previos de otras rondas (FR-1.18).
      required: [articuloId, nombre, unidadEsperada]
      properties:
        articuloId: { type: string, format: uuid }
        nombre: { type: string }
        codigo: { type: string, nullable: true }
        unidadEsperada: { $ref: '#/components/schemas/Unidad' }
        orden: { type: integer }
    Unidad:
      type: object
      required: [id, nombre, esPeso]
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        esPeso: { type: boolean }
    RegistroEntrada:
      type: object
      required: [articuloId, estado, claveIdempotencia, modoCaptura, capturadoEn, origenParse]
      properties:
        articuloId: { type: string, format: uuid }
        estado: { $ref: '#/components/schemas/EstadoRegistro' }
        cantidad:
          type: number
          minimum: 0
          nullable: true
          description: Obligatorio salvo `no_contado`. Solo números exactos (FR-1.9).
        unidadId: { type: string, format: uuid, nullable: true }
        modoCaptura: { $ref: '#/components/schemas/ModoCaptura' }
        origenParse: { type: string, enum: [gramatica, modelo, manual] }
        capturadoEn:
          type: string
          format: date-time
          description: Reloj del cliente anclado a la referencia del servidor (D-16).
        claveIdempotencia: { type: string, format: uuid }
        confirmaCorreccion:
          type: boolean
          default: false
          description: Requerido si el artículo ya fue contado en esta ronda (FR-1.14).
        confirmaPeseAAlerta:
          type: boolean
          default: false
          description: El usuario confirma pese a una alerta de discrepancia (FR-2.4).
    RegistroAceptado:
      type: object
      required: [registroId, secuencia, recibidoEn]
      properties:
        registroId: { type: string, format: uuid }
        secuencia: { type: integer }
        recibidoEn: { type: string, format: date-time }
...
    Error:
      type: object
      required: [codigo, mensaje]
      properties:
        codigo: { type: string }
        mensaje: { type: string }
        detalles: { type: object, additionalProperties: true }
  responses:
...
    Prohibido:
      description: El rol no autoriza esta operación
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflicto:
      description: Conflicto con el estado actual
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    EntradaInvalida:
      description: Falló la validación de esquema
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...